requirements.txt
africastalking==1.2.5
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
billiard==4.1.0
black==21.7b0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
contextlib2==21.6.0
Django==4.2.1
django-appconf==1.0.5
django-ckeditor==6.5.1
django-filter==23.2
django-imagekit==4.1.0
django-js-asset==2.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
flake8==3.9.2
geographiclib==2.0
geopy==2.3.0
idna==3.4
isort==5.9.2
kombu==5.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.8.1
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2021.1
regex==2021.7.6
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
tomli==1.0.4
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
